Develop

+++ src/iam/src/module/auth/features/v1/logout/logout.ts
@@ -1,17 +1,24 @@
 import { RabbitRPC } from '@golevelup/nestjs-rabbitmq';
-import { Inject, Injectable, Logger, NotFoundException } from '@nestjs/common';
+import {
+  ForbiddenException,
+  Inject,
+  Injectable,
+  Logger,
+  UnauthorizedException,
+} from '@nestjs/common';
 import { EventEmitter2 } from '@nestjs/event-emitter';
 import configs from 'building-blocks/configs/configs';
 import { EVENT_AUTH } from 'building-blocks/constants/event.constant';
 import { RoutingKey } from 'building-blocks/constants/rabbitmq.constant';
+import { RedisCacheService } from 'building-blocks/redis/redis-cache.service';
 import { randomQueueName } from 'building-blocks/utils/random-queue';
 import { IAuthRepository } from '../../../../../data/repositories/auth.repository';
-import { TokenType } from '../../../enums/token-type.enum';
-import { RedisCacheService } from 'building-blocks/redis/redis-cache.service';
 import { IUserRepository } from '../../../../../data/repositories/user.repository';
+import { TokenType } from '../../../enums/token-type.enum';
 
 export class Logout {
   accessToken: string;
+  userId: string;
 
   constructor(request: Partial<Logout> = {}) {
     Object.assign(this, request);
@@ -42,7 +49,11 @@ export class LogoutHandler {
       );
 
       if (!token) {
-        throw new NotFoundException('Token not found');
+        throw new UnauthorizedException('Token invalid');
+      }
+
+      if (token.userId != command.userId) {
+        throw new ForbiddenException('You can only logout of your account');
       }
 
       const tokenEntity = await this.authRepository.removeToken(token);
